move collection_keys to client
